fix: ft_nbrlen_base, ft_itoa_base data type fix

+++ ft_itoa_base.c
@@ -6,13 +6,13 @@
 /*   By: hyeongki <hyeongki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/26 17:33:35 by hyeongki          #+#    #+#             */
-/*   Updated: 2022/05/26 18:05:33 by hyeongki         ###   ########.fr       */
+/*   Updated: 2022/05/31 16:44:06 by hyeongki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_itoa_base(int n, int base)
+char	*ft_itoa_base(long long n, int base)
 {
 	int		i;
 	char	*ret;
